requirements.txt
beautifulsoup4
pytest-playwright
